+++ src/track/mod.rs
@@ -3,8 +3,7 @@ mod sample;
 mod attributes;
 mod offset;
 
-pub use track::Track;
-pub use track::TrackIdentifier;
+pub use track::{Track, TrackIdentifier, ParsableTrackId};
 pub use attributes::TrackAttributes;
 pub use offset::{SampleOffsets, SampleOffset};
 pub use sample::Sample;